refactor(hindsight): venues are data; netting is the marked fallback
Venue decoders are deleted. A venue is now only its address-book
section: entry points and fee collectors feed the shared netting
fallback and the fee bookkeeping; solver aliases feed attribution.
One netting module replaces the three (sender, venue, intent arms
picked by the entry point). CoW settlements decode from the Trade log
as declared data. TraderRole, DecodeContext, TradeDecoder, and the
receipt-only matching module are deleted — matching now also accepts
any transaction whose trace contains a known solver frame. Fee-wallet
venue attribution checks wallets in address order, so a trade cut by
two venues' wallets resolves the same way on every run.
