+++ tools/hindsight/src/decoder/attribution.rs
@@ -1,0 +1,595 @@
+//! Attribution: which solver settled a decoded trade, and which venue owns its order flow.
+//!
+//! Attribution runs after decoding and only labels the record (plus the fee bookkeeping a
+//! fee-wallet match implies). Nothing here affects whether a trade decodes.
+//!
+//! The solver label comes from the first evidence tier that answers, most- to least-trusted
+//! (see `AttributionSource`). The venue label is normally the contract the trader entered
+//! through (`tx.to`); some venues own the order flow without being that contract and are
+//! recognized from registry-driven fingerprints (owner, `appData` tag, fee wallet, integrator
+//! tag).
+
+use std::collections::HashSet;
+
+use alloy::{
+    primitives::{Address, B256, U256},
+    rpc::types::trace::geth::CallFrame,
+    sol,
+    sol_types::SolCall,
+};
+use serde::Serialize;
+
+use crate::decoder::{
+    netting::TraderFlow, registry::Registry, trace, transfer_ledger::TransferLedger,
+};
+
+/// The evidence tier that produced a record's solver label, most- to least-trusted.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[serde(rename_all = "snake_case")]
+pub(crate) enum AttributionSource {
+    /// The venue's entry calldata names the solver (`MetaMask`'s `aggregatorId`).
+    Declared,
+    /// The entry point (`tx.to`) is itself a known solver router: the trade settled there.
+    EntryPoint,
+    /// A known solver router was called inside the trace (venue-wrapped entries).
+    TraceMatch,
+    /// No known router anywhere: best guess is the external call that moved the most native
+    /// value (an unknown router's address).
+    LargestCall,
+    /// Even the guess was indeterminate (e.g. a token→token trace where no call moves value).
+    /// The record is labeled with its entry point — typically the venue's name — flagging it
+    /// for registry expansion.
+    Fallback,
+}
+
+/// A solver label and the evidence tier it came from.
+pub(crate) struct Attribution {
+    pub solver: String,
+    pub source: AttributionSource,
+}
+
+/// Attribute the solver that settled a matched transaction.
+///
+/// `declared` is the venue calldata's own claim (see `venue_declared_solver`), which outranks
+/// everything; the remaining tiers read the trace, ending at the entry-point label as the honest
+/// "don't know".
+pub(crate) fn solver(
+    declared: Option<String>,
+    root: &CallFrame,
+    entry_point: Address,
+    sender: Address,
+    registry: &Registry,
+) -> Attribution {
+    if let Some(solver) = declared {
+        return Attribution { solver, source: AttributionSource::Declared };
+    }
+    if registry.is_solver(entry_point) {
+        return Attribution {
+            solver: registry.label(entry_point),
+            source: AttributionSource::EntryPoint,
+        };
+    }
+    if let Some(found) = trace::find_solver_frame(root, registry).and_then(|frame| frame.to) {
+        return Attribution { solver: registry.label(found), source: AttributionSource::TraceMatch };
+    }
+    if let Some(guess) = trace::largest_external_call(root, entry_point, sender, registry) {
+        return Attribution {
+            solver: registry.label(guess),
+            source: AttributionSource::LargestCall,
+        };
+    }
+    Attribution { solver: registry.label(entry_point), source: AttributionSource::Fallback }
+}
+
+sol! {
+    /// The `MetaMask` Swap Router entry point (selector `0x5f575529`): `aggregatorId` names the
+    /// solver API that produced the route.
+    function swap(string aggregatorId, address tokenFrom, uint256 amount, bytes data);
+}
+
+/// The solver the venue's entry calldata declares, normalized to the address book's solver names
+/// via the venue's `solver_aliases` section. `None` when the entry point is not a venue that
+/// declares its solver, or the calldata is not the declaring call.
+///
+/// `MetaMask` states which solver API it routed through (e.g. "oneInchV6FeeDynamic",
+/// "uniswapPermit2FeeDynamic"). Trace attribution often cannot resolve these — a token→token
+/// route moves no native value and enters through Permit2 — so the calldata declaration is the
+/// authoritative source.
+pub(crate) fn venue_declared_solver(
+    registry: &Registry,
+    entry_point: Address,
+    input: &[u8],
+) -> Option<String> {
+    let venue = registry
+        .venue_name(entry_point)
+        .and_then(|name| registry.venue(name))
+        .filter(|venue| venue.declares_solver())?;
+    let call = swapCall::abi_decode(input).ok()?;
+    Some(venue.normalize_solver(&call.aggregatorId))
+}
+
+/// The order-flow venue for a decoded flow, when a fingerprint matches — overriding the
+/// entry-point label. Every fingerprint is registry-driven; nothing here knows about a specific
+/// venue or provider.
+///
+/// Four fingerprints, tried in order: owning trader (`[venue_owners]`), `CoW` `appData` tag
+/// (`[venue_appdata]`; the hash is extracted by the caller), fee wallet (`[venue_fees]`),
+/// provider integrator tag (`[venue_integrators]`; extracted by the caller).
+///
+/// On a fee-wallet match the fee lands on the flow. For netted amounts it is backed out —
+/// added back to the output or netted out of the input, whichever side it was taken from. For
+/// declared amounts (`amounts_are_declared`), the sides differ: an input-side fee is recorded
+/// only (the declared `amount_in` is read after the fee left), but an output-side fee is still
+/// grossed back — a fee-wallet venue's wallet is paid from the routing path directly, so the
+/// declared recipient's receipt is short of the swap's gross output by exactly the fee.
+pub(crate) fn venue(
+    registry: &Registry,
+    flow: &mut TraderFlow,
+    ledger: &TransferLedger,
+    integrator: Option<&str>,
+    app_data: Option<B256>,
+    amounts_are_declared: bool,
+) -> Option<String> {
+    if let Some(venue) = registry.venue_for_owner(flow.tracked) {
+        return Some(venue.to_string());
+    }
+    if let Some(venue) = app_data.and_then(|hash| registry.venue_for_appdata(hash)) {
+        return Some(venue.to_string());
+    }
+    if let Some((venue, fee)) = fee_venue(registry, ledger, flow.swap.token_in, flow.swap.token_out)
+    {
+        match (fee, amounts_are_declared) {
+            (VenueFee::Input(amount), false) => flow.net_input_fee(amount),
+            (VenueFee::Input(amount), true) => {
+                flow.venue_fee_in = flow.venue_fee_in.or(Some(amount));
+            }
+            (VenueFee::Output(amount), _) => flow.gross_output_fee(amount),
+        }
+        return Some(venue);
+    }
+    integrator
+        .and_then(|tag| registry.venue_for_integrator(tag))
+        .map(str::to_string)
+}
+
+/// Which side of the swap a venue took its fee from, with the amount.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+enum VenueFee {
+    /// Skimmed off the input before the swap, so the settled input is smaller than the user spent.
+    Input(U256),
+    /// Taken out of the output after the swap, so the settled output is larger than the user kept.
+    Output(U256),
+}
+
+/// The venue whose fee wallet took a cut of this trade, and which side it came from. `None` when no
+/// venue fee wallet received a non-zero amount of either swap token.
+///
+/// Both sides are checked because venues split on this: Phantom and Robinhood take the buy token,
+/// while Coinbase's Base App skims the sell token before routing. The output side is tried first —
+/// a wallet that received both tokens is being paid its cut in the token the user bought. The
+/// wallets are checked in address order, so two venues' wallets both taking a cut of one trade
+/// resolve to the same venue on every run.
+fn fee_venue(
+    registry: &Registry,
+    ledger: &TransferLedger,
+    token_in: Address,
+    token_out: Address,
+) -> Option<(String, VenueFee)> {
+    for (wallet, venue) in registry.venue_fees() {
+        let received = ledger.received_by(&HashSet::from([*wallet]));
+        let non_zero = |token: &Address| {
+            received
+                .get(token)
+                .copied()
+                .filter(|amount| !amount.is_zero())
+        };
+        if let Some(fee) = non_zero(&token_out) {
+            return Some((venue.clone(), VenueFee::Output(fee)));
+        }
+        if let Some(fee) = non_zero(&token_in) {
+            return Some((venue.clone(), VenueFee::Input(fee)));
+        }
+    }
+    None
+}
+
+#[cfg(test)]
+mod tests {
+    use alloy::primitives::{address, b256, Bytes};
+    use tycho_simulation::tycho_common::models::Chain;
+
+    use super::*;
+    use crate::decoder::test_utils::{addr, frame, make_transfer_log, swap, PERMIT2};
+
+    #[test]
+    fn test_declared_solver_with_trace_candidate() {
+        // MetaMask declares its solver in calldata; even a known router in the trace must not
+        // override it.
+        let registry = Registry::ethereum();
+        let oneinch = address!("0x111111125421ca6dc452d289314280a0f8842a65");
+        let mut root = frame("CALL", addr(1), addr(2), 0);
+        root.calls = vec![frame("CALL", addr(2), oneinch, 1000)];
+
+        let attribution = solver(Some("uniswap".to_string()), &root, addr(2), addr(1), &registry);
+        assert_eq!(attribution.solver, "uniswap");
+        assert_eq!(attribution.source, AttributionSource::Declared);
+    }
+
+    #[test]
+    fn test_direct_swap_entry_point() {
+        let registry = Registry::ethereum();
+        let oneinch = address!("0x111111125421ca6dc452d289314280a0f8842a65");
+        let root = frame("CALL", addr(1), oneinch, 0);
+
+        let attribution = solver(None, &root, oneinch, addr(1), &registry);
+        assert_eq!(attribution.solver, "1inch");
+        assert_eq!(attribution.source, AttributionSource::EntryPoint);
+    }
+
+    #[test]
+    fn test_relay_internal_solver() {
+        // Mirrors the real Relay tx: the client router calls 0x's AllowanceHolder.
+        // root(relay) -> [ relay (self-call), 0x AllowanceHolder (the solver) ]
+        let registry = Registry::ethereum();
+        let sender = addr(1);
+        let relay = address!("0xf5042e6ffac5a625d4e7848e0b01373d8eb9e222");
+        let zerox = address!("0x0000000000001ff3684f28c67538d4d072c22734");
+
+        let mut root = frame("CALL", sender, relay, 0);
+        root.calls = vec![frame("CALL", relay, relay, 0), frame("CALL", relay, zerox, 1000)];
+
+        let attribution = solver(None, &root, relay, sender, &registry);
+        assert_eq!(attribution.solver, "0x");
+        assert_eq!(attribution.source, AttributionSource::TraceMatch);
+    }
+
+    #[test]
+    fn test_relay_tycho_router() {
+        // Real tx 0x8b461c…: Relay ApprovalProxy -> Relay router -> Tycho router.
+        // The settling solver is Tycho even though it sits two levels deep.
+        let registry = Registry::ethereum();
+        let sender = addr(1);
+        let relay_proxy = address!("0xccc88a9d1b4ed6b0eaba998850414b24f1c315be");
+        let relay_router = address!("0xb92fe925dc43a0ecde6c8b1a2709c170ec4fff4f");
+        let tycho = address!("0x1f8db310f32d48b6180ff902ec60c586128cef47");
+
+        let mut router_call = frame("CALL", relay_proxy, relay_router, 0);
+        router_call.calls = vec![frame("CALL", relay_router, tycho, 0)];
+        let mut root = frame("CALL", sender, relay_proxy, 0);
+        root.calls = vec![router_call];
+
+        let attribution = solver(None, &root, relay_proxy, sender, &registry);
+        assert_eq!(attribution.solver, "tycho");
+        assert_eq!(attribution.source, AttributionSource::TraceMatch);
+    }
+
+    #[test]
+    fn test_unknown_solver_largest_external_call() {
+        // No known solver in the trace: pick the largest external call,
+        // skipping the client self-call and the refund back to the sender.
+        let registry = Registry::ethereum();
+        let sender = addr(1);
+        let client = addr(2);
+        let unknown_router = addr(50);
+
+        let mut root = frame("CALL", sender, client, 0);
+        root.calls = vec![
+            frame("CALL", client, client, 0),            // self-call, skipped
+            frame("CALL", client, sender, 9000),         // refund to sender, skipped
+            frame("CALL", client, addr(51), 10),         // small external call
+            frame("CALL", client, unknown_router, 5000), // largest external call
+        ];
+
+        let attribution = solver(None, &root, client, sender, &registry);
+        assert_eq!(attribution.solver, unknown_router.to_string());
+        assert_eq!(attribution.source, AttributionSource::LargestCall);
+    }
+
+    #[test]
+    fn test_attribution_zero_value_fallback() {
+        // Unknown solver, token->token swap: every child call moves zero value, so the guess
+        // would degenerate to the first child (the Permit2 token pull). The record is labeled
+        // with its entry point instead, marked as a fallback.
+        let registry = Registry::ethereum();
+        let sender = addr(1);
+        let client = addr(2);
+
+        let mut root = frame("CALL", sender, client, 0);
+        root.calls = vec![
+            frame("CALL", client, PERMIT2, 0),  // token pull
+            frame("CALL", client, addr(50), 0), // unknown solver, zero value
+        ];
+
+        let attribution = solver(None, &root, client, sender, &registry);
+        assert_eq!(attribution.solver, client.to_string());
+        assert_eq!(attribution.source, AttributionSource::Fallback);
+    }
+
+    #[test]
+    fn test_attribution_wrapped_native_frames() {
+        // ETH-input swap through an unknown router: the highest-value direct call is the
+        // WETH.deposit() wrapping the input. Infrastructure, not a solver — the guess must
+        // fall through to the real router call.
+        let registry = Registry::ethereum();
+        let sender = addr(1);
+        let client = addr(2);
+        let unknown = addr(50);
+
+        let mut root = frame("CALL", sender, client, 0);
+        root.calls = vec![
+            frame("CALL", client, registry.wrapped_native(), 9000), // wrap, skipped
+            frame("CALL", client, unknown, 100),
+        ];
+
+        let attribution = solver(None, &root, client, sender, &registry);
+        assert_eq!(attribution.solver, unknown.to_string());
+        assert_eq!(attribution.source, AttributionSource::LargestCall);
+    }
+
+    #[test]
+    fn test_attribution_permit2_frames() {
+        // Even when Permit2 is the highest-value direct call, it is infrastructure, not a solver.
+        let registry = Registry::ethereum();
+        let sender = addr(1);
+        let client = addr(2);
+        let unknown = addr(50);
+
+        let mut root = frame("CALL", sender, client, 0);
+        root.calls =
+            vec![frame("CALL", client, PERMIT2, 9000), frame("CALL", client, unknown, 100)];
+
+        let attribution = solver(None, &root, client, sender, &registry);
+        assert_eq!(attribution.solver, unknown.to_string());
+        assert_eq!(attribution.source, AttributionSource::LargestCall);
+    }
+
+    fn metamask_router(registry: &Registry) -> Address {
+        *registry
+            .venue("metamask")
+            .unwrap()
+            .entry_points
+            .iter()
+            .next()
+            .unwrap()
+    }
+
+    #[test]
+    fn test_venue_declared_solver_known_ids() {
+        let registry = Registry::ethereum();
+        let router = metamask_router(&registry);
+        for (id, want) in [
+            ("oneInchV6FeeDynamic", "1inch"),
+            ("uniswapPermit2FeeDynamic", "uniswap"),
+            ("okx6", "okx"),
+            ("someFutureSolver", "someFutureSolver"),
+        ] {
+            let call = swapCall {
+                aggregatorId: id.to_string(),
+                tokenFrom: addr(10),
+                amount: U256::from(1000),
+                data: Bytes::default(),
+            };
+            assert_eq!(
+                venue_declared_solver(&registry, router, &call.abi_encode()).as_deref(),
+                Some(want)
+            );
+        }
+    }
+
+    #[test]
+    fn test_venue_declared_solver_other_selectors_and_venues() {
+        let registry = Registry::ethereum();
+        let router = metamask_router(&registry);
+        // Another selector on the declaring venue: no declaration.
+        assert_eq!(venue_declared_solver(&registry, router, &[0xde, 0xad, 0xbe, 0xef, 0x00]), None);
+        assert_eq!(venue_declared_solver(&registry, router, &[]), None);
+        // A venue with no solver aliases (Relay) never declares, even with matching calldata.
+        let relay = address!("0xb92fe925dc43a0ecde6c8b1a2709c170ec4fff4f");
+        let call = swapCall {
+            aggregatorId: "oneInchV6FeeDynamic".to_string(),
+            tokenFrom: addr(10),
+            amount: U256::from(1000),
+            data: Bytes::default(),
+        };
+        assert_eq!(venue_declared_solver(&registry, relay, &call.abi_encode()), None);
+    }
+
+    #[test]
+    fn test_attributes_owner_to_venue() {
+        // A CoW-settled kpk trade nets to the Safe that owns the order; the venue is that Safe.
+        let registry = Registry::ethereum();
+        let kpk_safe = address!("0x4f2083f5fbede34c2714affb3105539775f7fe64");
+        let ledger = TransferLedger::from_transaction(&[], &[]);
+        let mut flow = TraderFlow::without_fees(kpk_safe, swap(addr(10), 1, addr(11), 2));
+        assert_eq!(venue(&registry, &mut flow, &ledger, None, None, false).as_deref(), Some("kpk"));
+    }
+
+    #[test]
+    fn test_unknown_owner_is_not_a_venue() {
+        let registry = Registry::ethereum();
+        let ledger = TransferLedger::from_transaction(&[], &[]);
+        let mut flow = TraderFlow::without_fees(addr(9), swap(addr(10), 1, addr(11), 2));
+        assert_eq!(venue(&registry, &mut flow, &ledger, None, None, false), None);
+    }
+
+    #[test]
+    fn test_appdata_tag_attributes_venue() {
+        // A CoW order carrying DefiLlama's appData hash is attributed to LlamaSwap; an unregistered
+        // hash is not.
+        let registry = Registry::ethereum();
+        let ledger = TransferLedger::from_transaction(&[], &[]);
+        let defillama = b256!("0xf249b3db926aa5b5a1b18f3fec86b9cc99b9a8a99ad7e8034242d2838ae97422");
+        let mut flow = TraderFlow::without_fees(addr(1), swap(addr(10), 1, addr(11), 2));
+        assert_eq!(
+            venue(&registry, &mut flow, &ledger, None, Some(defillama), false).as_deref(),
+            Some("llamaswap")
+        );
+        assert_eq!(venue(&registry, &mut flow, &ledger, None, Some(B256::ZERO), false), None);
+    }
+
+    #[test]
+    fn test_fee_wallet_attributes_and_grosses_fee_back() {
+        // A 0x-routed Phantom swap: the buy-token fee reaches Phantom's wallet. It must be added
+        // back so the settled output is gross (else every Phantom swap under-reports by 85 bps).
+        let registry = Registry::ethereum();
+        let phantom = address!("0x2cffed5d56eb6a17662756ca0fdf350e732c9818");
+        let user = addr(1);
+        let pool = addr(50);
+        let token_in = addr(10);
+        let token_out = addr(11);
+        let logs = vec![
+            make_transfer_log(token_in, user, pool, U256::from(1000)),
+            make_transfer_log(token_out, pool, user, U256::from(9915)),
+            make_transfer_log(token_out, pool, phantom, U256::from(85)),
+        ];
+        let ledger = TransferLedger::from_transaction(&logs, &[]);
+        let mut flow = TraderFlow::without_fees(user, swap(token_in, 1000, token_out, 9915));
+
+        assert_eq!(
+            venue(&registry, &mut flow, &ledger, None, None, false).as_deref(),
+            Some("phantom")
+        );
+        assert_eq!(flow.venue_fee_out, Some(U256::from(85)));
+        assert_eq!(flow.swap.amount_out, U256::from(10000));
+    }
+
+    #[test]
+    fn test_fee_wallet_on_declared_amounts() {
+        // The same Phantom fee leg on a declared decode: the wallet is paid from the routing
+        // path directly, so the declared recipient's receipt (9915) is short of the gross output
+        // by the fee — grossed back, exactly like a netted flow.
+        let registry = Registry::ethereum();
+        let phantom = address!("0x2cffed5d56eb6a17662756ca0fdf350e732c9818");
+        let user = addr(1);
+        let pool = addr(50);
+        let token_in = addr(10);
+        let token_out = addr(11);
+        let logs = vec![
+            make_transfer_log(token_in, user, pool, U256::from(1000)),
+            make_transfer_log(token_out, pool, user, U256::from(9915)),
+            make_transfer_log(token_out, pool, phantom, U256::from(85)),
+        ];
+        let ledger = TransferLedger::from_transaction(&logs, &[]);
+        let mut flow = TraderFlow::without_fees(user, swap(token_in, 1000, token_out, 9915));
+
+        assert_eq!(
+            venue(&registry, &mut flow, &ledger, None, None, true).as_deref(),
+            Some("phantom")
+        );
+        assert_eq!(flow.venue_fee_out, Some(U256::from(85)));
+        assert_eq!(flow.swap.amount_out, U256::from(10000));
+    }
+
+    #[test]
+    fn test_fee_wallet_input_fee_on_declared_amounts_is_recorded_only() {
+        // An input-side wallet fee on a declared decode: the declared amount_in was read from
+        // the solver frame, after the fee left — netting it out again would double-subtract.
+        let registry = Registry::builtin(Chain::Bsc).unwrap();
+        let coinbase = address!("0x5aafc1f252d544f744d17a4e734afd6efc47ede4");
+        let user = addr(1);
+        let pool = addr(50);
+        let token_in = addr(10);
+        let token_out = addr(11);
+        let logs = vec![
+            make_transfer_log(token_in, user, coinbase, U256::from(95)),
+            make_transfer_log(token_in, user, pool, U256::from(9905)),
+            make_transfer_log(token_out, pool, user, U256::from(2000)),
+        ];
+        let ledger = TransferLedger::from_transaction(&logs, &[]);
+        let mut flow = TraderFlow::without_fees(user, swap(token_in, 9905, token_out, 2000));
+
+        assert_eq!(
+            venue(&registry, &mut flow, &ledger, None, None, true).as_deref(),
+            Some("coinbase")
+        );
+        assert_eq!(flow.venue_fee_in, Some(U256::from(95)));
+        assert_eq!(flow.swap.amount_in, U256::from(9905));
+    }
+
+    #[test]
+    fn test_integrator_tag_attributes_venue() {
+        // A provider integrator tag maps to its venue, case-insensitively; an unknown tag does
+        // not.
+        let registry = Registry::ethereum();
+        let ledger = TransferLedger::from_transaction(&[], &[]);
+        let mut flow = TraderFlow::without_fees(addr(1), swap(addr(10), 1, addr(11), 2));
+        assert_eq!(
+            venue(&registry, &mut flow, &ledger, Some("Infinex"), None, false).as_deref(),
+            Some("infinex")
+        );
+        assert_eq!(venue(&registry, &mut flow, &ledger, Some("somedapp"), None, false), None);
+    }
+
+    #[test]
+    fn test_fee_wallet_input_side_fee_nets_the_input_down() {
+        // A LiFi-routed Coinbase Base App swap: the 0.95% cut is skimmed off the sell token before
+        // routing, so only the remainder reached the pools. Leaving it in makes the settled trade
+        // look bigger than it was and Fynd, re-solved on that inflated size, appear to win.
+        let registry = Registry::builtin(Chain::Bsc).unwrap();
+        let coinbase = address!("0x5aafc1f252d544f744d17a4e734afd6efc47ede4");
+        let user = addr(1);
+        let pool = addr(50);
+        let token_in = addr(10);
+        let token_out = addr(11);
+        let logs = vec![
+            make_transfer_log(token_in, user, coinbase, U256::from(95)),
+            make_transfer_log(token_in, user, pool, U256::from(9905)),
+            make_transfer_log(token_out, pool, user, U256::from(2000)),
+        ];
+        let ledger = TransferLedger::from_transaction(&logs, &[]);
+        let mut flow = TraderFlow::without_fees(user, swap(token_in, 10000, token_out, 2000));
+
+        assert_eq!(
+            venue(&registry, &mut flow, &ledger, Some("base-app"), None, false).as_deref(),
+            Some("coinbase")
+        );
+        assert_eq!(flow.venue_fee_in, Some(U256::from(95)));
+        assert_eq!(flow.swap.amount_in, U256::from(9905));
+        // The output side is untouched: this venue took nothing out of the buy token.
+        assert_eq!(flow.venue_fee_out, None);
+        assert_eq!(flow.swap.amount_out, U256::from(2000));
+    }
+
+    #[test]
+    fn test_fee_wallet_taking_both_tokens_is_read_as_an_output_fee() {
+        // A wallet that received both swap tokens is being paid its cut in the token the user
+        // bought; the sell-token leg is the swap's own routing, not a second fee.
+        let registry = Registry::ethereum();
+        let phantom = address!("0x2cffed5d56eb6a17662756ca0fdf350e732c9818");
+        let user = addr(1);
+        let token_in = addr(10);
+        let token_out = addr(11);
+        let logs = vec![
+            make_transfer_log(token_in, user, phantom, U256::from(7)),
+            make_transfer_log(token_out, addr(50), phantom, U256::from(85)),
+        ];
+        let ledger = TransferLedger::from_transaction(&logs, &[]);
+        let mut flow = TraderFlow::without_fees(user, swap(token_in, 1000, token_out, 9915));
+
+        assert_eq!(
+            venue(&registry, &mut flow, &ledger, None, None, false).as_deref(),
+            Some("phantom")
+        );
+        assert_eq!(flow.venue_fee_out, Some(U256::from(85)));
+        assert_eq!(flow.swap.amount_out, U256::from(10000));
+        assert_eq!(flow.venue_fee_in, None);
+        assert_eq!(flow.swap.amount_in, U256::from(1000));
+    }
+
+    #[test]
+    fn test_no_fee_transfer_is_not_a_venue() {
+        // Dust to the fee wallet in a token other than the output is not this trade's fee.
+        let registry = Registry::ethereum();
+        let user = addr(1);
+        let pool = addr(50);
+        let token_in = addr(10);
+        let token_out = addr(11);
+        let logs = vec![
+            make_transfer_log(token_in, user, pool, U256::from(1000)),
+            make_transfer_log(token_out, pool, user, U256::from(2000)),
+        ];
+        let ledger = TransferLedger::from_transaction(&logs, &[]);
+        let mut flow = TraderFlow::without_fees(user, swap(token_in, 1000, token_out, 2000));
+        assert_eq!(venue(&registry, &mut flow, &ledger, None, None, false), None);
+    }
+}
