+++ tools/hindsight/src/decoder/transfer_ledger.rs
@@ -178,56 +178,6 @@ impl TransferLedger {
             .fold(U256::ZERO, |total, &(_, _, _, value)| total.saturating_add(value))
     }
 
-    /// Gross total received per token by any of `recipients`, regardless of sender (native ETH
-    /// keyed by `Address::ZERO`).
-    pub(crate) fn received_by(&self, recipients: &HashSet<Address>) -> HashMap<Address, U256> {
-        let mut totals: HashMap<Address, U256> = HashMap::new();
-        if recipients.is_empty() {
-            return totals;
-        }
-        for &(token, _, to, value) in &self.transfers {
-            if recipients.contains(&to) {
-                *totals.entry(token).or_default() += value;
-            }
-        }
-        totals
-    }
-
-    /// Per-token net outflow of the address group: what the group sent minus what it got back,
-    /// where positive.
-    pub(crate) fn group_net_sent(&self, group: &HashSet<Address>) -> HashMap<Address, U256> {
-        let (sent, received) = self.group_totals(group);
-        net_positive(&sent, &received)
-    }
-
-    /// Per-token net inflow of the address group: what the group received minus what it sent,
-    /// where positive.
-    pub(crate) fn group_net_received(&self, group: &HashSet<Address>) -> HashMap<Address, U256> {
-        let (sent, received) = self.group_totals(group);
-        net_positive(&received, &sent)
-    }
-
-    /// Aggregated receipts of **pure sinks** — addresses that received value but never sent any
-    /// in the transaction — as `(recipient, token, total)`. A pool or router always sends
-    /// something back, so a pure sink is a delivery endpoint (an order's recipient, a payout).
-    pub(crate) fn sink_receipts(&self) -> Vec<(Address, Address, U256)> {
-        let mut senders: HashSet<Address> = HashSet::new();
-        for &(_, from, _, _) in &self.transfers {
-            senders.insert(from);
-        }
-        let mut received: HashMap<(Address, Address), U256> = HashMap::new();
-        for &(token, _, to, value) in &self.transfers {
-            if !senders.contains(&to) {
-                *received.entry((to, token)).or_default() += value;
-            }
-        }
-        received
-            .into_iter()
-            .filter(|(_, total)| !total.is_zero())
-            .map(|((recipient, token), total)| (recipient, token, total))
-            .collect()
-    }
-
     /// Totals of `token` paid to pure sinks — addresses that received value in the transaction
     /// but never sent any — by senders that never received the token themselves: a trader
     /// spending it or a pool paying out a swap, never a router forwarding what it was paid.
@@ -253,53 +203,20 @@ impl TransferLedger {
             .collect()
     }
 
-    /// Gross flow of one token through the transaction, counting every transfer of it in either
-    /// direction. The denominator of the residue test.
-    pub(crate) fn token_gross(&self, token: Address) -> U256 {
-        let mut gross = U256::ZERO;
-        for &(transferred, _, _, value) in &self.transfers {
-            if transferred == token {
-                gross = gross.saturating_add(value);
+    /// Gross total received per token by any of `recipients`, regardless of sender (native ETH
+    /// keyed by `Address::ZERO`).
+    pub(crate) fn received_by(&self, recipients: &HashSet<Address>) -> HashMap<Address, U256> {
+        let mut totals: HashMap<Address, U256> = HashMap::new();
+        if recipients.is_empty() {
+            return totals;
+        }
+        for &(token, _, to, value) in &self.transfers {
+            if recipients.contains(&to) {
+                *totals.entry(token).or_default() += value;
             }
         }
-        gross
+        totals
     }
-
-    /// Gross sent and received per token, summed across the group.
-    fn group_totals(
-        &self,
-        group: &HashSet<Address>,
-    ) -> (HashMap<Address, U256>, HashMap<Address, U256>) {
-        let mut sent: HashMap<Address, U256> = HashMap::new();
-        let mut received: HashMap<Address, U256> = HashMap::new();
-        for &(token, from, to, value) in &self.transfers {
-            if group.contains(&from) {
-                *sent.entry(token).or_default() += value;
-            }
-            if group.contains(&to) {
-                *received.entry(token).or_default() += value;
-            }
-        }
-        (sent, received)
-    }
-}
-
-/// Per-token `positive - negative` where positive, over the union of tokens.
-fn net_positive(
-    positive: &HashMap<Address, U256>,
-    negative: &HashMap<Address, U256>,
-) -> HashMap<Address, U256> {
-    let mut net: HashMap<Address, U256> = HashMap::new();
-    for (&token, &amount) in positive {
-        let offset = negative
-            .get(&token)
-            .copied()
-            .unwrap_or_default();
-        if amount > offset {
-            net.insert(token, amount - offset);
-        }
-    }
-    net
 }
 
 /// A net leg is residue when its token routed between third parties and the leg is under this
@@ -758,64 +675,5 @@ mod tests {
             U256::from(999u64)
         );
         assert_eq!(transfer_ledger.received_by_address(addr(200), token), U256::ZERO);
-    }
-
-    #[test]
-    fn test_group_round_trips() {
-        // The group sends 1000 of token A and gets 300 back: net sent 700, no net receipt.
-        let group = HashSet::from([addr(99)]);
-        let logs = vec![
-            make_transfer_log(addr(10), addr(99), addr(50), U256::from(1000)),
-            make_transfer_log(addr(10), addr(50), addr(99), U256::from(300)),
-            make_transfer_log(addr(11), addr(50), addr(99), U256::from(200)),
-        ];
-        let transfer_ledger = TransferLedger::from_transaction(&logs, &[]);
-        assert_eq!(
-            transfer_ledger.group_net_sent(&group),
-            HashMap::from([(addr(10), U256::from(700))])
-        );
-        assert_eq!(
-            transfer_ledger.group_net_received(&group),
-            HashMap::from([(addr(11), U256::from(200))])
-        );
-    }
-
-    #[test]
-    fn test_sink_payments() {
-        // The trader pays token_a to the pool (a sender, so never a sink) and, split by the
-        // token contract, to a wallet that only accumulates. On the token_b side the pool is
-        // the source of the fee wallet's leg, while the router only forwards what it received.
-        let trader = addr(1);
-        let pool = addr(50);
-        let router = addr(2);
-        let in_fee_wallet = addr(60);
-        let out_fee_wallet = addr(61);
-        let token_a = addr(10);
-        let token_b = addr(11);
-
-        let logs = vec![
-            make_transfer_log(token_a, trader, pool, U256::from(950)),
-            make_transfer_log(token_a, trader, in_fee_wallet, U256::from(30)),
-            make_transfer_log(token_a, trader, in_fee_wallet, U256::from(20)),
-            make_transfer_log(token_b, pool, trader, U256::from(1900)),
-            make_transfer_log(token_b, pool, out_fee_wallet, U256::from(100)),
-            make_transfer_log(token_b, pool, router, U256::from(500)),
-            make_transfer_log(token_b, router, out_fee_wallet, U256::from(500)),
-        ];
-        let transfer_ledger = TransferLedger::from_transaction(&logs, &[]);
-
-        assert_eq!(transfer_ledger.sink_payments(token_a), vec![(in_fee_wallet, U256::from(50))]);
-        assert_eq!(transfer_ledger.sink_payments(token_b), vec![(out_fee_wallet, U256::from(100))]);
-    }
-
-    #[test]
-    fn test_sink_receipts_when_recipient_also_sent() {
-        // The pool receives and sends (a conversion), the recipient only receives (a sink).
-        let logs = vec![
-            make_transfer_log(addr(10), addr(1), addr(50), U256::from(1000)),
-            make_transfer_log(addr(11), addr(50), addr(7), U256::from(2000)),
-        ];
-        let transfer_ledger = TransferLedger::from_transaction(&logs, &[]);
-        assert_eq!(transfer_ledger.sink_receipts(), vec![(addr(7), addr(11), U256::from(2000))]);
     }
 }
